[ 박수로 ] 리팩토링 : 캘린더컴포넌트 중복 코드 제거
