+++ src/index.js
@@ -9,26 +9,28 @@ import App from './App';
 import createSagaMiddleware from 'redux-saga';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import axios from 'axios';
-
+import { errorMessage } from './functions/errorHandling';
 const createStoreWithMiddleware = applyMiddleware(
   PromiseMiddleware,
   ReduxThunk,
 )(createStore);
 
 axios.defaults.withCredentials = true;
-
+// axios.interceptors.response.use(
+//   function (response) {
+//     return response;
+//   },
+//   function (error) {
+//     errorMessage(error.response?.data.message);
+//     return Promise.reject(error);
+//   },
+// );
 ReactDOM.render(
-  // <Provider
-  //   store={createStoreWithMiddleware(
-  //     Reducer,
-  //     composeWithDevTools(applyMiddleware(sagaMiddleware)),
-  //   )}
-  // >
   <Provider
-     store={createStoreWithMiddleware(
-       Reducer,
-       composeWithDevTools(applyMiddleware()),
-     )}
+    store={createStoreWithMiddleware(
+      Reducer,
+      composeWithDevTools(applyMiddleware()),
+    )}
   >
     <App />
   </Provider>,
